move: dto 경로 정리 (#18)
