+++ src/main/java/modelly/modelly_be/domain/recruitment/dto/common/RecruitmentSchedule.java
@@ -1,4 +1,4 @@
-package modelly.modelly_be.domain.recruitment.dto;
+package modelly.modelly_be.domain.recruitment.dto.common;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
