+++ src/main/java/modelly/modelly_be/domain/recruitment/dto/request/RecruitmentRequestDto.java
@@ -1,8 +1,9 @@
-package modelly.modelly_be.domain.recruitment.dto;
+package modelly.modelly_be.domain.recruitment.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+import modelly.modelly_be.domain.recruitment.dto.common.RecruitmentSchedule;
 import modelly.modelly_be.global.entity.Category;
 
 import java.util.List;
